requirements.txt
Pillow
python-dotenv
coloredlogs
requests